refactor: extract shared code to packages
