+++ apps/www/package.json
@@ -29,6 +29,8 @@
     "@tpluscode/rdf-ns-builders": "^2.0.1",
     "@tpluscode/rdf-string": "^0.2.26",
     "@tpluscode/sparql-builder": "^0.3.23",
+    "@view-builder/core": "0.0.0",
+    "@view-builder/view-util": "0.0.0",
     "@zazuko/vocabulary-extras": "^1.2.3",
     "alcaeus": "^2.3.2",
     "clownface": "^1.5.1",
@@ -37,7 +39,6 @@
     "is-graph-pointer": "^1.2.2",
     "lit": "^2.3.0",
     "n3": "^1.16.2",
-    "rdf-cube-view-query": "^1.8.2",
     "rdf-ext": "^2.0.1",
     "sparql-http-client": "^2.4.1",
     "string-to-stream": "^3.0.1",
