+++ apps/www/test/automation.test.js
@@ -2,8 +2,8 @@ import { expect } from 'chai'
 import sinon from 'sinon'
 import $rdf from 'rdf-ext'
 import { rdfs, rdf } from '@tpluscode/rdf-ns-builders'
+import { ssz, view } from '@view-builder/core/ns.js'
 import { generateDimensions } from '../src/automation.js'
-import { ssz, view } from '../src/ns.js'
 import { ex, testData } from './_helpers.js'
 
 describe('automation.js', () => {
