+++ packages/core/ns.js
