+++ packages/core/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "@view-builder/core",
+  "version": "0.0.0",
+  "type": "module",
+  "dependencies": {
+    "@rdfjs/namespace": "^2.0.0",
+    "@zazuko/vocabulary-extras": "^1.2.3"
+  }
+}
