+++ packages/view-util/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "@view-builder/view-util",
+  "version": "0.0.0",
+  "type": "module",
+  "main": "index.js",
+  "dependencies": {
+    "@rdfjs/dataset": "^2.0.0",
+    "@view-builder/core": "0.0.0",
+    "clownface": "^1.5.1",
+    "rdf-cube-view-query": "^1.10.0"
+  }
+}
